Introducing namespace 'DTC' for all classes loaded by Load.C
Prime intention is to Use a namespace DTC for the class 'Layer' in order
to avoid a conflict in the Root 6 dictionary generation.

When compiling with Root 6, a class named 'Layer' is found in addition
to the class 'Layer' of the DTCToolkid. The exact reason is not yet
understood. The duplicate candidate is 'TMVA::DNN::Layer' from Root
include/TMVA/NeuralNet

A separate namespace for the DTC code is anyhow advisable.